Move openUrl intent to main source set
